% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvEBayes_object_S3_methods.R
\name{BIC.pvEBayes}
\alias{BIC.pvEBayes}
\title{Obtain Bayesian Information Criterion (BIC) for a pvEBayes object}
\usage{
\method{BIC}{pvEBayes}(object, ...)
}
\arguments{
\item{object}{a \code{pvEBayes} object, which is the output of the function
\link{pvEBayes} or \link{pvEBayes_tune}.}

\item{...}{other input parameters. Currently unused.}
}
\value{
numeric, BIC score for the resulting model.
}
\description{
This function defines the S3 \code{BIC} method for objects of class
\code{pvEBayes}. It extracts the Bayesian Information Criterion (BIC)
from a fitted model.
}
\examples{

fit <- pvEBayes(
  contin_table = statin2025_44, model = "general-gamma",
  alpha = 0.3, n_posterior_draws = NULL
)

BIC_score <- BIC(fit)

}
