% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbca2025_69}
\alias{gbca2025_69}
\title{FDA GBCA dataset with 69 adverse events}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 70 rows and 8 columns.
}
\source{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
\usage{
gbca2025_69
}
\description{
An adverse event-drug count dataset (contingency table)
obtained from the FDA FAERS database for the
quarters 2021Q1 - 2024Q4
}
\details{
Data are stored in the form of a contingency table, with
drugs listed across the columns and the 69 AEs presented across
the rows. Each cell in the contingency table represents the total
report counts associated with that (AE, drug) pair and detected
in the FDA FAERS database during 2021Q1 - 2024Q4.

The dataset catalogs 7 Gadolinium-Based Contrast Agents (GBCAs)
(across columns):

Gadobenate, Gadobutrol, Gadodiamide, Gadopentetate, Gadoterate, Gadoteridol,
Gadoxetate.

The 69 adverse events presented across the rows are selected from 1328 AEs of
gbca2025 which are related to the brain or neural system. Other AEs are
collapsed to one reference row: "Other AEs".
}
\keyword{datasets}
