% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvEBayes_object_S3_methods.R
\name{print.pvEBayes}
\alias{print.pvEBayes}
\title{Print method for a pvEBayes object}
\usage{
\method{print}{pvEBayes}(x, ...)
}
\arguments{
\item{x}{a \code{pvEBayes} object, which is the output of the function
\link{pvEBayes} or \link{pvEBayes_tune}.}

\item{...}{other input parameters. Currently unused.}
}
\value{
Invisibly returns the input \code{pvEBayes} object.
}
\description{
This function defines the S3 \code{print} method for objects of class
\code{pvEBayes}. It displays a concise description of the fitted model.
}
\examples{

obj <- pvEBayes(
  contin_table = statin2025_44, model = "general-gamma",
  alpha = 0.5, n_posterior_draws = 10000
)

print(obj)
}
