% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statin2025_44}
\alias{statin2025_44}
\title{FDA statin dataset with 44 adverse events}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 45 rows and 7 columns.
}
\source{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
\usage{
statin2025_44
}
\description{
An adverse event-drug count dataset (contingency table)
obtained from the FDA FAERS database for the
quarters 2021Q1 - 2024Q4.
}
\details{
Data are stored in the form of a contingency table, with
drugs listed across the columns and the 44 AEs presented across
the rows. Each cell in the contingency table represents the total
report counts associated with that (AE, drug) pair and detected
in the FDA FAERS database during 2021Q1 - 2024Q4.

The dataset catalogs 6 statin drugs (across columns):

Atorvastatin, Fluvastatin, Lovastatin, Pravastatin, Rosuvastatin, Simvastatin.

The 44 adverse events presented across the rows are considered
significant by FDA.

This dataset is an updated version of statin46 from the pvLRT package which
collect the same scope of AEs for 6 statin drugs for quarters 2014Q3 - 2020Q4.

During 2021Q1 - 2024Q4, there was no AE report for "BLOOD CREATINE
PHOSPHOKINASE MM INCREASED" and "MYOGLOBIN BLOOD PRESENT". Therefore, these
two AEs are not presented in the statin2025_44 dataset.
}
\keyword{datasets}
