% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvEBayes_object_S3_methods.R
\name{summary_table_pvEBayes}
\alias{summary_table_pvEBayes}
\title{Obtain a summary table for a pvEBayes object}
\usage{
summary_table_pvEBayes(x, cutoff_signal = 1.001)
}
\arguments{
\item{x}{a \code{pvEBayes} object, which is the output of the function
\link{pvEBayes} or \link{pvEBayes_tune}.}

\item{cutoff_signal}{numeric. Threshold for signal detection. An AE-drug
combination is classified as a detected signal if its 5th posterior
percentile exceeds this threshold.}
}
\value{
a data.table that summarizes reporting count (N), expected null
value (E), posterior probability of being a signal (post_prob), posterior
signal strength median (q50), 5-th and 95-th posterior signal strength
percentile (q05 and q95) for each AE-drug combination.
}
\description{
Obtain a summary table for a pvEBayes object
}
\examples{

fit <- pvEBayes(
  contin_table = statin2025_44, model = "general-gamma",
  alpha = 0.5, n_posterior_draws = 100
)

summary_table_pvEBayes(fit)

}
