% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vigi_opioid_mental}
\alias{vigi_opioid_mental}
\title{VigiBase opioid dataset with 100 mental-related adverse events}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 101 rows and 6 columns.
}
\source{
\url{https://www.vigiaccess.org/}
}
\usage{
vigi_opioid_mental
}
\description{
An adverse event-drug count dataset (contingency table)
obtained from VigiBase.
}
\details{
Data are stored in the form of a contingency table, with
drugs listed across the columns and the 100 AEs presented across
the rows. Each cell in the contingency table represents the total
report counts associated with that (AE, drug) pair and detected
in VigiBase.

The dataset catalogs 5 opioid drugs (across columns):

Codeine, Fentanyl, Oxycodone, Pentazocine and Tramadol.
}
\references{
Tan Y, Markatou M and Chakraborty S. A Review of Statistical Methods for
Spontaneous Reporting System Data Mining: Signal Detection and Beyond.
\emph{arXiv}:2604.18898 (stat.AP). https://doi.org/10.48550/arXiv.2604.18898
}
\keyword{datasets}
