% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvEBayes_object_S3_methods.R
\name{summary.pvEBayes}
\alias{summary.pvEBayes}
\title{Summary method for a pvEBayes object}
\usage{
\method{summary}{pvEBayes}(object, return = NULL, ...)
}
\arguments{
\item{object}{a \code{pvEBayes} object, which is the output of the function
\link{pvEBayes} or \link{pvEBayes_tune}.}

\item{return}{a character string specifying which component the summary
function should return.Valid options include: "prior parameters",
"likelihood", "detected signal" and "posterior draws". If set to NULL
(default), a summary table will be returned (see 'summary_table_pvEBayes()').
Note that the input for 'return' is case-sensitive.}

\item{...}{other input parameters. Currently unused.}
}
\value{
If \code{return = NULL} (default), the function returns a summary table generated
by \code{summary_table_pvEBayes()}, after printing the fitted \code{pvEBayes} object.

If \code{return} is specified, the function returns the requested component:
\describe{
\item{\verb{prior parameters}}{A list of estimated prior parameters.}
\item{\code{likelihood}}{The fitted model log marginal likelihood.}
\item{\verb{detected signal}}{A logical matrix indicating AE-drug pairs with
posterior signal probability at least 0.95.}
\item{\verb{posterior draws}}{Posterior draws of the signal strength for each
AE-drug pair. }
}
}
\description{
This function defines the S3 \code{summary} method for objects of class
\code{pvEBayes}. It provides a detailed summary of the fitted model.
}
\examples{

obj <- pvEBayes(
  contin_table = statin2025_44, model = "general-gamma",
  alpha = 0.5, n_posterior_draws = 10000
)

summary(obj)

}
